split rag retrieval into multiple files
